+++ include/pfnet/parser.h
@@ -23,7 +23,7 @@ typedef struct Parser Parser;
 // Prototypes
 Parser* PARSER_new(void);
 Parser* PARSER_new_for_file(char* f);
-void PARSER_init(Parser* p);
+void PARSER_init(Parser* p, BOOL params);
 Net* PARSER_parse(Parser* p, char* f, int num_periods);
 void PARSER_set(Parser* p, char* key, REAL value);
 void PARSER_show(Parser* p);
@@ -37,12 +37,12 @@ void* PARSER_get_data(Parser* p);
 void PARSER_set_error(Parser* p, char* string);
 
 void PARSER_set_data(Parser* p, void* data);
-void PARSER_set_func_init(Parser* p, void (*func)(Parser* p));
+void PARSER_set_func_init(Parser* p, void (*func)(Parser* p, BOOL init_params));
 void PARSER_set_func_parse(Parser* p, Net* (*func)(Parser* p, char* f, int n));
 void PARSER_set_func_set(Parser* p, void (*func)(Parser* p, char* key, REAL v));
 void PARSER_set_func_show(Parser* p, void (*func)(Parser* p));
 void PARSER_set_func_write(Parser* p, void (*func)(Parser* p, Net* net, char* f));
-void PARSER_set_func_free(Parser* p, void (*func)(Parser* p));
+void PARSER_set_func_free(Parser* p, void (*func)(Parser* p, BOOL del_parser));
 
 
 #endif