+++ include/pfnet/parser_JSON.h
@@ -21,12 +21,12 @@ typedef struct JSON_Parser JSON_Parser;
 
 // Interface
 Parser* JSON_PARSER_new(void);
-void JSON_PARSER_init(Parser* parser);
-Net* JSON_PARSER_parse(Parser* parser, char* filename, int num_periods);
-void JSON_PARSER_set(Parser* parser, char* key, REAL value);
-void JSON_PARSER_show(Parser* parser);
-void JSON_PARSER_write(Parser* parser, Net* net, char* filename);
-void JSON_PARSER_free(Parser* parser);
+void JSON_PARSER_init(Parser* p, BOOL init_params);
+Net* JSON_PARSER_parse(Parser* p, char* filename, int num_periods);
+void JSON_PARSER_set(Parser* p, char* key, REAL value);
+void JSON_PARSER_show(Parser* p);
+void JSON_PARSER_write(Parser* p, Net* net, char* filename);
+void JSON_PARSER_free(Parser* p, BOOL del_parser);
 
 // Others
 void JSON_PARSER_process_json_bus_array(Parser* p, Net* net, json_value* json_bus_array);