+++ include/pfnet/parser_MAT.h
@@ -56,12 +56,12 @@ typedef struct MAT_Util MAT_Util;
 
 // Interface
 Parser* MAT_PARSER_new(void);
-void MAT_PARSER_init(Parser* p);
+void MAT_PARSER_init(Parser* p, BOOL init_params);
 Net* MAT_PARSER_parse(Parser* p, char* f, int num_periods);
 void MAT_PARSER_set(Parser* p, char* key, REAL value);
 void MAT_PARSER_show(Parser* p);
 void MAT_PARSER_write(Parser* p, Net* net, char* f);
-void MAT_PARSER_free(Parser* p);
+void MAT_PARSER_free(Parser* p, BOOL del_parser);
 
 // MAT-specific
 void MAT_PARSER_load(MAT_Parser* p, Net* net);