+++ src/parser/parser_MAT.c
@@ -121,18 +121,23 @@ Parser* MAT_PARSER_new(void) {
   PARSER_set_func_show(p,&MAT_PARSER_show);
   PARSER_set_func_write(p,&MAT_PARSER_write);
   PARSER_set_func_free(p,&MAT_PARSER_free);
-  PARSER_init(p);
+  PARSER_init(p,TRUE);
   return p;
 }
 
-void MAT_PARSER_init(Parser* p) {
+void MAT_PARSER_init(Parser* p, BOOL init_params) {
 
+  // Local variables
+  MAT_Parser* parser;
+  
   // No parser
   if (!p)
     return;
 
   // Allocate
-  MAT_Parser* parser = (MAT_Parser*)malloc(sizeof(MAT_Parser));
+  parser = (MAT_Parser*)PARSER_get_data(p);
+  if (!parser)
+    parser = (MAT_Parser*)malloc(sizeof(MAT_Parser));
   
   // Error
   parser->error_flag = FALSE;
@@ -144,8 +149,10 @@ void MAT_PARSER_init(Parser* p) {
   parser->record = 0;
   MAT_PARSER_clear_token(parser);
   
-  // Options
-  parser->output_level = 0;
+  // Parameters
+  if (init_params) {
+    parser->output_level = 0;
+  }
 
   // Base
   parser->base_power = MAT_PARSER_BASE_POWER;
@@ -304,7 +311,7 @@ void MAT_PARSER_write(Parser* p, Net* net, char* f) {
   // nothing
 }
 
-void MAT_PARSER_free(Parser* p) {
+void MAT_PARSER_free(Parser* p, BOOL del_parser) {
 
   // Local variables
   MAT_Parser* parser = (MAT_Parser*)PARSER_get_data(p);
@@ -331,7 +338,8 @@ void MAT_PARSER_free(Parser* p) {
   LIST_map(MAT_Util,parser->util_list,util,next,{free(util);});
 
   // Free parser
-  free(parser);
+  if (del_parser)
+    free(parser);
 }
 
 void MAT_PARSER_load(MAT_Parser* parser, Net* net) {